Thêm các hàm tính phí cho Phòng và Sinh Viên

+++ Entities/DICHVU.cs
@@ -9,9 +9,9 @@ namespace KTX_Management.Entities
     class DICHVU
     {
         // Fields
-        public double GiaDV1;
-        public double GiaDV2;
-        public double GiaDV3;
+        public int GiaDV1;
+        public int GiaDV2;
+        public int GiaDV3;
         public static DICHVU DVChung = new DICHVU(320000, 30000);
         public static DICHVU DVCaNhan = new DICHVU(20000, 60000, 50000);
         // Methods
@@ -19,12 +19,12 @@ namespace KTX_Management.Entities
         {
 
         }
-        public DICHVU(double GiaDV1, double GiaDV2)
+        public DICHVU(int GiaDV1, int GiaDV2)
         {
             this.GiaDV1 = GiaDV1;
             this.GiaDV2 = GiaDV2;
         }
-        public DICHVU(double GiaDV1, double GiaDV2, double GiaDV3)
+        public DICHVU(int GiaDV1,int GiaDV2, int GiaDV3)
         {
             this.GiaDV1 = GiaDV1;
             this.GiaDV2 = GiaDV2;
